+++ app/upgrades/v7_5/constants.go
@@ -1,0 +1,25 @@
+package v75
+
+import (
+	store "cosmossdk.io/store/types"
+	circuittypes "cosmossdk.io/x/circuit/types"
+
+	"github.com/kiichain/kiichain/v7/app/upgrades"
+)
+
+const (
+	// UpgradeName is the name of the upgrade
+	UpgradeName = "v7.5.0"
+)
+
+// Upgrade defines the coordinated upgrade that adds the Cosmos SDK x/circuit
+// module store so msg-type circuit breakers can be enabled on-chain.
+var Upgrade = upgrades.Upgrade{
+	UpgradeName:          UpgradeName,
+	CreateUpgradeHandler: CreateUpgradeHandler,
+	StoreUpgrades: store.StoreUpgrades{
+		Added: []string{
+			circuittypes.StoreKey,
+		},
+	},
+}